Fork the devember project to be omphalos proper
This change forks (sort of) the devember 2022 project into a proper
project repository.

The repo isn't an exact fork though, since you cannot fork a repo from
your account to your account and I'm not interested in putting this
particular project in an org repo (at least, for the moment).

As such, this is just a duplicate of the repo which shares a history
up to but not including this commit, which changes the state of a few
things for our purposes.

1. Docs are now deployed in a separate repo, and so the docs folder here
   is what the website folder used to be; a github action deploys the
   docs to that repository automagically when a push happens.

2. The CNAME changed so that the documentation ends up where we expect
   it to.

3. Some of the doc content changed for link reasons.

4. The README has changed since this is no longer a devember project.