Use Svelte stores to update connection counts
This does the groundwork needed for having the list of graphics display
a live count of the number of active connections.

There are two new stores; one that provides live updates of connection
status, and one that derives from that and augments a list of static
graphics with the current connection count (which defaults to 0 when
there are no active connections).