Force all panels to have an explicit workspace
We're working through some changes to the dashboard, and to do so we
need all of the panels to have a defined workspace name.

This field is optional, so we now force in the appropriate default value
when it is missing (which was always planned but we never got to it
during the initial development).