Coalesce the dashboard name and slug into one
At least for the forseeable future, the information on workspaces now
presumes that the name and the slug are the same thing, and so the
structure of the object chained.

This is how this works in NodeCG (more or less; they use a different URL
format and lower case things, but otherwise the same), and I see no
compelling reason currently to alter that, since it would require more
metadata changes.