Track the current workspace in the navbar
This change brings in the oft desired but up until now not present
because I was too dense to realize how quick the fix was feature of
having the current dashboard tab be marked as active.

We moved the parent router location to encompass the entire root page
instead of just the content area in the center, which allows all
components on the page to know about the route in use, which makes the
entire thing super simple to implement.