Reload the dashboard on workspace change
This is a small change, but it allows us to actually cause a dashboard
component refresh to happen when the page route changes.

This allows a panel change to occur so that not only does the active tab
show what the active workspace is, but also (in theory) the actual
panels for it display too.

To support this, the panel information that we gather from the bundle
information we're given is given information on the workspace each of
the panels is in, so that they can be displayed in the appropriate
place.