Update the navbar tab style and apply everywhere
Previously only the workspaces had a tab style, and they had a mark at
all times and not just when they were active (though the style
changed).

We want all buttons that navigate to local pages to appear active when
they are, but not when they are not since the split in the inactive tab
underline between the left and right looks weird.

So, a simple new tab component that can make itself be active or not has
been added and this is used for everything except the docs link
(which is not a tab).

The component is being used because it's unclear to me if it's possible
to capture in a function call the element whose attributes are
changing, and if so how to do that.