Change the documentation link for developer mode
For the purposes of local testing and development, it makes more sense
to link to the current in progress docs than the officially posted
ones, since we likely want to see and refer to changes we're currently
working on.

To that end, the documentation link changes where it goes depending on
the developer mode used.