Include new API methods for adding custom routers
With the upcoming goal being to try and run obotnurd within the omphalos
code, there is a missing bit of API functionality that is strictly
required, which is the ability to set up custom routes.

The code for obotnurd's TTS and file editing components require routes
to tie things into the back end code.