Change the dashboard panel row height
The previous setting for panel heights in the dashboard was set to cause
the height of rows to make for square panels (i.e. each row was the
same height as the columns are wide).

Now that we're using this in a real world context, that is far too large
and so the size has been constrained down to allow for more
granularity.