Fix an issue with client side message init
The protocol that the clients use to talk to each other and the server
requires that the first message sent be a `HELO` message.

There was an issue in that the `HELO` is triggered when the socket
connects, but the user initialization code might want to ship a message
right away.

This changes the ordering of things so that the announcement happens
right away.