Get rid of explicit HELO messages in the protocol
Previously all sockets needed to explicitly announce who they were on
connect with a HELO message because the server rejects all messages from
a socket that it doesn't have address information on.

This is an issue with reconnections because while disconnected messages
are queued and will be sent once the connection re-establishes itself,
which would cause them all to be ignored because HELO has not happened
yet.

This change tightens the protocol by having the auth information that is
given to the server at connect time the appropriate HELO information,
allowing it to implicitly know right at connect who the other end is.