Get rid of explicit JOIN on connect
Previously the code would, on connect and on reconnect, explicitly JOIN
to the bundle room. An issue with this is that if any listens were in
progress for messages from other bundles, those JOIN would not be done
which causes message drops.

This change updates both mechanisms so that part of the new auth that is
sent is the list of rooms to immediately join to. Initially this will
be just the bundle room, but once listens from other bundles happen it
will include those too.

JOIN/PART that happen while disconnected will still work because those
messages will be queued for sending, and will be received once the
connection re-establishes, bringing things back into the correct state.