Fix exceptions on URL copy on insecure sites
During the initial first deployment of the bot proper, we discovered
that copying the URL doesnt work because the hostname isn't localhost
and the protocol isn't https (due to internal dev host).

That's fine for us, but to ensure that this does not confuse others,
this situation is now caught and an error toast is displayed so that
you know why exactly it's not doing what you expect it to.