Support serving sound files from bundles
As the first step in setting up our soundboard support, this includes
the core logic for validating the information in the bundle as relates
to sounds, as well as the code required to set up routes for sounds and
serve them.

With this in place, it is possible to request sounds from bundles and
have them play in the browser. Front end work is also needed for this,
which will be next.