Include some initial mixer page changes
This change introduces a new change to the mixer page that will show all
of the sounds across all bundles, as well as a sound selection drop box
for the available sound outputs.

This is still a little rough, pending the completion of the sound API,
at which point as we cycle back through we can clean everything up.